Refactoring the android code structure. Removed references to Fragment obsolete.
