+++ android/src/main/kotlin/top/kikt/imagescanner/core/ResultHandler.kt
@@ -1,4 +1,4 @@
-package top.kikt.imagescanner
+package top.kikt.imagescanner.core
 
 
 import android.os.Handler
