+++ android/src/main/kotlin/top/kikt/imagescanner/core/permission/PermissionsListener.java
@@ -1,4 +1,4 @@
-package top.kikt.imagescanner;
+package top.kikt.imagescanner.core.permission;
 
 /**
  * Created by Administrator on 2017/2/19.
