+++ android/src/main/kotlin/top/kikt/imagescanner/core/refresh/RefreshObserver.kt
@@ -1,4 +1,4 @@
-package top.kikt.imagescanner
+package top.kikt.imagescanner.core.refresh
 
 import android.database.ContentObserver
 import android.os.Handler
